requirements.txt
apache-beam[gcp]==2.54.0
mysql-connector-python==8.0.33
google-cloud-bigquery==3.11.4